// src/permission/permission.enum.ts
export enum Permissions {
  createUser = 'create user',
  readUser = 'read user',
  updateUser = 'update user',
  deleteUser = 'delete user',
  createRole = 'create role',
  readRole = 'read role',
  updateRole = 'update role',
  deleteRole = 'delete role',
  createPermission = 'create permission',
  deletePermission = 'delete permission',
  readPermission = 'read permission',
  updatePermission = 'update permission',
}